requirements.txt
CherryPy==18.9.0
firebase_admin==6.3.0
more_itertools==10.1.0
pydantic==2.5.2
pytest==7.4.3
python-dotenv==1.0.0
python-telegram-bot==20.7
pytest-cov==4.1.0
pre-commit==3.5.0
coverage==7.3.2
certifi==2023.11.17
annotated-types==0.6.0
Requests==2.31.0
spotipy==2.23.0
typing_extensions==4.8.0
pytest-asyncio==0.23.3
git+https://github.com/heisen273/ytmusicapi.git@ef0837efdd4d769dd6c3294eac986150aa756254
